+++ src/consumer/component/dependencies-resolver.js
@@ -1,7 +1,7 @@
 /** @flow */
 import path from 'path';
 import R from 'ramda';
-import { DEFAULT_INDEX_NAME, COMPONENT_ORIGINS } from '../../constants';
+import { COMPONENT_ORIGINS } from '../../constants';
 import ComponentMap from '../bit-map/component-map';
 import { BitId } from '../../bit-id';
 import Component from '../component';
@@ -9,7 +9,8 @@ import { Driver } from '../../driver';
 import { pathNormalizeToLinux, pathRelative, pathJoinLinux } from '../../utils';
 import logger from '../../logger/logger';
 import { Consumer } from '../../consumer';
-import type { RelativePath } from './consumer-component';
+import type { RelativePath } from './dependencies/dependency';
+import { Dependencies } from './dependencies';
 
 /**
  * Given the tree of file dependencies from the driver, find the components of these files.
@@ -23,6 +24,7 @@ import type { RelativePath } from './consumer-component';
  *
  * @param {Object} tree - contains direct deps for each file
  * @param {string[]} files - component files to search
+ * @param {string[]} testsFiles - component test files
  * @param {string} entryComponentId - component id for the entry of traversing - used to know which of the files are part of that component
  * @param {consumer} consumer
  * @param {string} bindingPrefix
@@ -32,14 +34,17 @@ import type { RelativePath } from './consumer-component';
 function findComponentsOfDepsFiles(
   tree: Object,
   files: string[],
+  testsFiles: string[],
   entryComponentId: string,
   consumer: Consumer,
   bindingPrefix: string,
-  componentFromModel,
+  componentFromModel: Component,
   driver: Driver
 ): Object {
   const packagesDeps = {};
+  let devPackagesDeps = {};
   const componentsDeps = {};
+  let devComponentsDeps = {};
   const untrackedDeps = [];
   const relativeDeps = []; // dependencies that are required with relative path (and should be required using 'bit/').
   const missingDeps = [];
@@ -105,9 +110,9 @@ function findComponentsOfDepsFiles(
         // dependencies array this component with the relativePaths array. Find the relativePath of this dep-file
         // to get the correct destinationRelativePath. There is no other way to obtain this info.
         const componentBitId = BitId.parse(componentId);
-        const dependency = componentFromModel.component.dependencies.find(
-          dep => dep.id.toStringWithoutVersion() === componentBitId.toStringWithoutVersion()
-        );
+        const dependency = componentFromModel.component
+          .getAllDependencies()
+          .find(dep => dep.id.toStringWithoutVersion() === componentBitId.toStringWithoutVersion());
         if (!dependency) {
           throw new Error(
             `the auto-generated file ${depFile} should be connected to ${componentId}, however, it's not part of the model dependencies of ${
@@ -137,7 +142,12 @@ function findComponentsOfDepsFiles(
     return { componentId, depFileRelative, destination };
   };
 
-  const processDepFile = (depFile: string, importSpecifiers?: Object, linkFile?: string) => {
+  const processDepFile = (
+    depFile: string,
+    importSpecifiers?: Object,
+    linkFile?: string,
+    isTestFile: boolean = false
+  ) => {
     if (processedFiles.includes(depFile)) return;
     processedFiles.push(depFile);
 
@@ -191,32 +201,52 @@ function findComponentsOfDepsFiles(
     if (componentsDeps[componentId]) {
       // it is another file of an already existing component. Just add the new path
       componentsDeps[componentId].push(depsPaths);
+    } else if (devComponentsDeps[componentId]) {
+      devComponentsDeps[componentId].push(depsPaths);
+    } else if (isTestFile) {
+      Object.assign(devComponentsDeps, currentComponentsDeps);
     } else {
       Object.assign(componentsDeps, currentComponentsDeps);
     }
   };
 
-  const processLinkFile = (linkFile) => {
+  const processLinkFile = (linkFile: Object, isTestFile: boolean = false) => {
     if (!linkFile.dependencies || R.isEmpty(linkFile.dependencies)) return;
     linkFile.dependencies.forEach((dependency) => {
       const component = getComponentIdByDepFile(linkFile.file);
       if (component.componentId) {
         // the linkFile is already a component, no need to treat it differently than other depFile
-        processDepFile(linkFile.file, dependency.importSpecifiers);
+        processDepFile(linkFile.file, dependency.importSpecifiers, undefined, isTestFile);
       } else {
-        processDepFile(dependency.file, dependency.importSpecifiers, linkFile.file);
+        processDepFile(dependency.file, dependency.importSpecifiers, linkFile.file, isTestFile);
       }
     });
   };
 
-  files.forEach((file) => {
-    const currentPackagesDeps = tree[file].packages;
-    if (currentPackagesDeps && !R.isEmpty(currentPackagesDeps)) {
-      Object.assign(packagesDeps, currentPackagesDeps);
+  const processDepFiles = (allDepsFiles: string[], importSpecifiers: Object, isTestFile: boolean = false) => {
+    if (allDepsFiles && !R.isEmpty(allDepsFiles)) {
+      allDepsFiles.forEach((depFile) => {
+        let finalImportSpecifiers = R.clone(importSpecifiers);
+        if (importSpecifiers) {
+          const importSpecifiersFound = importSpecifiers.find(specifierFile => specifierFile.file === depFile);
+          if (importSpecifiersFound) finalImportSpecifiers = importSpecifiersFound.importSpecifiers;
+        }
+        processDepFile(depFile, finalImportSpecifiers, undefined, isTestFile);
+      });
     }
-    const currentBitsDeps = tree[file].bits;
-    if (currentBitsDeps && !R.isEmpty(currentBitsDeps)) {
-      currentBitsDeps.forEach((bitDep) => {
+  };
+
+  const processLinkFiles = (allLinkFiles: Object[], isTestFile: boolean = false) => {
+    if (allLinkFiles && !R.isEmpty(allLinkFiles)) {
+      allLinkFiles.forEach((linkFile) => {
+        processLinkFile(linkFile, isTestFile);
+      });
+    }
+  };
+
+  const processBits = (bits, isTestFile: boolean = false) => {
+    if (bits && !R.isEmpty(bits)) {
+      bits.forEach((bitDep) => {
         const componentId = Consumer.getComponentIdFromNodeModulesPath(bitDep, bindingPrefix);
         const getExistingId = () => {
           let existingId = consumer.bitMap.getExistingComponentId(componentId);
@@ -233,9 +263,9 @@ function findComponentsOfDepsFiles(
             }
           }
           if (componentFromModel) {
-            const modelDep = componentFromModel.dependencies.find(
-              dep => dep.id.toStringWithoutVersion() === componentId
-            );
+            const modelDep = componentFromModel
+              .getAllDependencies()
+              .find(dep => dep.id.toStringWithoutVersion() === componentId);
             if (modelDep) return modelDep.id.toString();
           }
           return null;
@@ -244,34 +274,51 @@ function findComponentsOfDepsFiles(
         if (existingId) {
           const currentComponentsDeps = { [existingId]: [] };
           if (!componentsDeps[existingId]) {
-            Object.assign(componentsDeps, currentComponentsDeps);
+            if (isTestFile) {
+              Object.assign(devComponentsDeps, currentComponentsDeps);
+            } else {
+              Object.assign(componentsDeps, currentComponentsDeps);
+            }
           }
         } else {
           missingDeps.push(componentId);
         }
       });
     }
-    const allDepsFiles = tree[file].files;
-    if (allDepsFiles && !R.isEmpty(allDepsFiles)) {
-      allDepsFiles.forEach((depFile) => {
-        let importSpecifiers;
-        if (tree[file].importSpecifiers) {
-          const importSpecifiersFound = tree[file].importSpecifiers.find(
-            specifierFile => specifierFile.file === depFile
-          );
-          if (importSpecifiersFound) importSpecifiers = importSpecifiersFound.importSpecifiers;
-        }
-        processDepFile(depFile, importSpecifiers);
-      });
+  };
+
+  const processPackages = (packages, isTestFile) => {
+    if (packages && !R.isEmpty(packages)) {
+      if (isTestFile) {
+        Object.assign(devPackagesDeps, packages);
+      } else {
+        Object.assign(packagesDeps, packages);
+      }
     }
-    const allLinkFiles = tree[file].linkFiles;
-    if (allLinkFiles && !R.isEmpty(allLinkFiles)) {
-      allLinkFiles.forEach((linkFile) => {
-        processLinkFile(linkFile);
-      });
-    }
+  };
+
+  /**
+   * Remove the dependencies which appear both in dev and regular deps from the dev
+   * Because if a dependency is both dev dependency and regular dependecy it should be treated as regular one
+   * Apply for both packages and components dependencies
+   */
+  const removeDevDepsIfTheyAlsoRegulars = () => {
+    const devPackagesOnlyNames = R.difference(R.keys(devPackagesDeps), R.keys(packagesDeps));
+    devPackagesDeps = R.pick(devPackagesOnlyNames, devPackagesDeps);
+    const devComponentsOnlyNames = R.difference(R.keys(devComponentsDeps), R.keys(componentsDeps));
+    devComponentsDeps = R.pick(devComponentsOnlyNames, devComponentsDeps);
+  };
+
+  files.forEach((file) => {
+    const isTestFile = R.contains(file, testsFiles);
+    processPackages(tree[file].packages, isTestFile);
+    processBits(tree[file].bits, isTestFile);
+    processDepFiles(tree[file].files, tree[file].importSpecifiers, isTestFile);
+    processLinkFiles(tree[file].linkFiles, isTestFile);
   });
-  return { componentsDeps, packagesDeps, untrackedDeps, relativeDeps, missingDeps };
+  removeDevDepsIfTheyAlsoRegulars();
+
+  return { componentsDeps, devComponentsDeps, packagesDeps, devPackagesDeps, untrackedDeps, relativeDeps, missingDeps };
 }
 
 /**
@@ -362,6 +409,7 @@ export default (async function loadDependenciesForComponent(
   const traversedDeps = findComponentsOfDepsFiles(
     dependenciesTree.tree,
     allFiles,
+    testsFiles,
     idWithConcreteVersionString,
     consumer,
     component.bindingPrefix,
@@ -369,12 +417,20 @@ export default (async function loadDependenciesForComponent(
     driver
   );
   const traversedCompDeps = traversedDeps.componentsDeps;
-  component.dependencies = Object.keys(traversedCompDeps).map((depId) => {
+  const traversedCompDevDeps = traversedDeps.devComponentsDeps;
+
+  const dependencies = Object.keys(traversedCompDeps).map((depId) => {
     return { id: BitId.parse(depId), relativePaths: traversedCompDeps[depId] };
   });
+  component.setDependencies(dependencies);
+  const devDependencies = Object.keys(traversedCompDevDeps).map((depId) => {
+    return { id: BitId.parse(depId), relativePaths: traversedCompDevDeps[depId] };
+  });
+  component.setDevDependencies(devDependencies);
   const untrackedDependencies = traversedDeps.untrackedDeps;
   if (!R.isEmpty(untrackedDependencies)) missingDependencies.untrackedDependencies = untrackedDependencies;
   component.packageDependencies = traversedDeps.packagesDeps;
+  component.devPackageDependencies = traversedDeps.devPackagesDeps;
   if (!R.isEmpty(traversedDeps.relativeDeps)) {
     missingDependencies.relativeComponents = traversedDeps.relativeDeps;
   }